Seperated customer favorites management from CustomerController following SRP
