+++ src/main/java/com/groovify/vinylshopapi/services/CustomerService.java
@@ -2,19 +2,15 @@ package com.groovify.vinylshopapi.services;
 
 import com.groovify.vinylshopapi.dtos.*;
 import com.groovify.vinylshopapi.enums.RoleType;
-import com.groovify.vinylshopapi.exceptions.ConflictException;
 import com.groovify.vinylshopapi.exceptions.RecordNotFoundException;
 import com.groovify.vinylshopapi.mappers.CustomerMapper;
 import com.groovify.vinylshopapi.mappers.OrderMapper;
-import com.groovify.vinylshopapi.mappers.VinylRecordMapper;
 import com.groovify.vinylshopapi.models.Customer;
 import com.groovify.vinylshopapi.models.Order;
 import com.groovify.vinylshopapi.models.Role;
-import com.groovify.vinylshopapi.models.VinylRecord;
 import com.groovify.vinylshopapi.repositories.CustomerRepository;
 import com.groovify.vinylshopapi.repositories.OrderRepository;
 import com.groovify.vinylshopapi.repositories.RoleRepository;
-import com.groovify.vinylshopapi.repositories.VinylRecordRepository;
 import com.groovify.vinylshopapi.specifications.CustomerSpecification;
 import com.groovify.vinylshopapi.utils.SortHelper;
 import com.groovify.vinylshopapi.validation.ValidationUtils;
@@ -35,8 +31,6 @@ public class CustomerService {
     private final OrderRepository orderRepository;
     private final OrderMapper orderMapper;
     private final RoleRepository roleRepository;
-    private final VinylRecordRepository vinylRecordRepository;
-    private final VinylRecordMapper vinylRecordMapper;
     private final ValidationUtils validationUtils;
     private final PasswordEncoder passwordEncoder;
 
@@ -46,8 +40,6 @@ public class CustomerService {
             OrderRepository orderRepository,
             OrderMapper orderMapper,
             RoleRepository roleRepository,
-            VinylRecordRepository vinylRecordRepository,
-            VinylRecordMapper vinylRecordMapper,
             ValidationUtils validationUtils,
             PasswordEncoder passwordEncoder
     ) {
@@ -56,8 +48,6 @@ public class CustomerService {
         this.orderRepository = orderRepository;
         this.orderMapper = orderMapper;
         this.roleRepository = roleRepository;
-        this.vinylRecordRepository = vinylRecordRepository;
-        this.vinylRecordMapper = vinylRecordMapper;
         this.validationUtils = validationUtils;
         this.passwordEncoder = passwordEncoder;
     }
@@ -118,35 +108,6 @@ public class CustomerService {
     }
 
 
-    public List<VinylRecordSummaryResponseDTO> getCustomerFavoriteRecords(Long customerId) {
-        List<VinylRecord> favoriteRecords = findCustomer(customerId).getFavoriteVinylRecords();
-        return vinylRecordMapper.toSummaryResponseDTOs(favoriteRecords);
-    }
-
-    public void addFavoriteRecordToCustomer(Long customerId, Long vinylRecordId) {
-        Customer customer = findCustomer(customerId);
-        VinylRecord favoriteRecord = findVinylRecord(vinylRecordId);
-
-        if (customer.getFavoriteVinylRecords().contains(favoriteRecord)) {
-            throw new ConflictException("Vinyl record with id: '" + vinylRecordId + "' is already in customer's favorite list.");
-        }
-
-        customer.getFavoriteVinylRecords().add(favoriteRecord);
-        customerRepository.save(customer);
-    }
-
-    public void removeFavoriteRecordFromCustomer(Long customerId, Long vinylRecordId) {
-        Customer customer = findCustomer(customerId);
-        VinylRecord favoriteRecord = findVinylRecord(vinylRecordId);
-
-        if (!customer.getFavoriteVinylRecords().contains(favoriteRecord)) {
-            throw new ConflictException("No vinyl record found in the customer's favorite list with id: " + vinylRecordId);
-        }
-
-        customer.getFavoriteVinylRecords().remove(favoriteRecord);
-        customerRepository.save(customer);
-    }
-
     public List<OrderSummaryResponseDTO> getCustomerOrders(Long customerId) {
         return orderMapper.toOrderSummaryResponseDTOs(findCustomer(customerId).getOrders());
     }
@@ -155,14 +116,10 @@ public class CustomerService {
         return orderMapper.toResponseDTO(findOrder(orderId, customerId));
     }
 
+
     private Customer findCustomer(Long customerId) {
         return customerRepository.findByIdAndIsDeletedFalse(customerId)
                 .orElseThrow(() -> new RecordNotFoundException("No customer found with id: " + customerId));
-    }
-
-    private VinylRecord findVinylRecord(Long vinylRecordId) {
-        return vinylRecordRepository.findById(vinylRecordId)
-                .orElseThrow(() -> new RecordNotFoundException("No vinyl record found with id: " + vinylRecordId));
     }
 
     private Order findOrder(Long orderId, Long customerId) {
